interp: skip kill goroutine if it's useless
If the context has no Done channel (such as context.Background), the
fired goroutine will simply block forever and waste resources.

This shows very clearly in the benchmark.

name   old time/op    new time/op    delta
Run-4    1.41ms ± 2%    1.27ms ± 0%  -10.28%  (p=0.002 n=6+6)

name   old alloc/op   new alloc/op   delta
Run-4     143kB ± 0%     142kB ± 0%   -0.49%  (p=0.004 n=5+6)

name   old allocs/op  new allocs/op  delta
Run-4       481 ± 0%       479 ± 0%   -0.42%  (p=0.026 n=5+6)

+++ interp/module.go
@@ -57,20 +57,22 @@ func DefaultExec(ctx Ctxt, path string, args []string) error {
 
 	err := cmd.Start()
 	if err == nil {
-		go func() {
-			<-ctx.Context.Done()
-
-			if ctx.KillTimeout <= 0 || runtime.GOOS == "windows" {
-				_ = cmd.Process.Signal(os.Kill)
-				return
-			}
-
+		if done := ctx.Context.Done(); done != nil {
 			go func() {
-				time.Sleep(ctx.KillTimeout)
-				_ = cmd.Process.Signal(os.Kill)
+				<-done
+
+				if ctx.KillTimeout <= 0 || runtime.GOOS == "windows" {
+					_ = cmd.Process.Signal(os.Kill)
+					return
+				}
+
+				go func() {
+					time.Sleep(ctx.KillTimeout)
+					_ = cmd.Process.Signal(os.Kill)
+				}()
+				_ = cmd.Process.Signal(os.Interrupt)
 			}()
-			_ = cmd.Process.Signal(os.Interrupt)
-		}()
+		}
 
 		err = cmd.Wait()
 	}
